requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
PyPDF2
python-docx
sentence-transformers
huggingface-hub>=0.24.0